requirements.txt
django>=4.2
python-dotenv
requests
gunicorn
django-allauth
PyJWT
cryptography